models2/research/cognitive_mapping_and_planning/requirements_carla.txt
Pillow
numpy
protobuf
pygame
matplotlib
future
configparser
numpy
h5py
scipy
scikit-image
opencv-python
pid_controller
pickle